Fix 32-bit targets and restore Windows/macOS --find-interpreter
- Cast st_size to i64 on 32-bit platforms (armv7/x86 have i32 st_size)
- Restore --find-interpreter for Windows and macOS (no PyPy there)
- Keep explicit CPython interpreter list for Linux/musllinux only

+++ src/lib.rs
@@ -1900,7 +1900,7 @@ fn batch_open_io(filenames: &[String], exts: &[&str]) -> Vec<(usize, Arc<PreSeri
                     libc::stat(c_names[i].as_ptr(), &mut stat_buf)
                 }
             };
-            if rc == 0 { stat_buf.st_size } else { -1 }
+            if rc == 0 { stat_buf.st_size as i64 } else { -1 }
         })
         .collect();
